requirements.txt
streamlit
python-dotenv
pdfplumber
matplotlib
langchain-groq
langchain-core
